+++ src/convertSchema.test.ts
@@ -20,6 +20,7 @@ describe('convertSchema', () => {
       },
       numberOfEmployees: {
         bsonType: 'string',
+        enum: ['1 - 5', '6 - 20', '21 - 50', '51 - 200', '201 - 500', '500+'],
       },
       addresses: {
         bsonType: 'array',
@@ -120,8 +121,7 @@ describe('convertSchema', () => {
           fields: { keyword: { type: 'keyword', ignore_above: 256 } },
         },
         numberOfEmployees: {
-          type: 'text',
-          fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+          type: 'keyword',
         },
         addresses: {
           properties: {
@@ -185,6 +185,88 @@ describe('convertSchema', () => {
           type: 'text',
           fields: { keyword: { type: 'keyword', ignore_above: 256 } },
         },
+      },
+    })
+  })
+  test('Convert MongoDB schema to Elastic with options', () => {
+    const options = {
+      omit: ['integrations', 'permissions'],
+      overrides: [
+        { path: 'addresses.address.latitude', bsonType: 'double' },
+        { path: 'addresses.address.longitude', bsonType: 'double' },
+      ],
+    }
+    expect(convertSchema(schema, options)).toEqual({
+      properties: {
+        parentId: { type: 'keyword' },
+        name: {
+          type: 'text',
+          fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+        },
+        subType: {
+          type: 'text',
+          fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+        },
+        numberOfEmployees: {
+          type: 'keyword',
+        },
+        addresses: {
+          properties: {
+            address: {
+              properties: {
+                address1: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                address2: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                city: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                county: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                state: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                zip: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                country: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+                latitude: { type: 'double' },
+                longitude: { type: 'double' },
+                timezone: {
+                  type: 'text',
+                  fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+                },
+              },
+            },
+            name: {
+              type: 'text',
+              fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+            },
+            isPrimary: { type: 'boolean' },
+          },
+        },
+        logo: {
+          type: 'text',
+          fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+        },
+        verified: { type: 'boolean' },
+        partner: {
+          type: 'text',
+          fields: { keyword: { type: 'keyword', ignore_above: 256 } },
+        },
+        createdAt: { type: 'date' },
       },
     })
   })